add mssql to reservation

+++ servers/reservation/models/room.go
@@ -2,10 +2,10 @@ package models
 
 // Room represents the room information
 type Room struct {
-	ID         int64  `json:"roomID"`
+	ID         int    `json:"roomID"`
 	Name       string `json:"roomName"`
-	Floor      int64  `json:"floor"`
-	Capacity   int64  `json:"capacity"`
+	Floor      int    `json:"floor"`
+	Capacity   int    `json:"capacity"`
 	RoomType   string `json:"roomType"`
 	StatusType string `json:"roomStatus"`
 }
